metadata-leak-tester /requirements.txt
Flask==2.3.2
Pillow==10.0.0
PyPDF2==3.0.1
pdfminer.six==20221105
python-dotenv==1.0.0
mysql-connector-python==8.0.33
weasyprint==58.0
exifread==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2
html5lib==1.1
cffi==1.15.1